渲染逻辑重构： -添加了RenderCommonBuffer类 - SceneRenderer可以实例化，不再是只有static方法的类 - 渲染器命令都写入RenderCommonBuffer对象 - 改造了opengl context的概念

+++ KeyEngine/src/Key/Renderer/Image.h
@@ -118,7 +118,7 @@ namespace Key {
 
 		inline uint32_t CalculateMipCount(uint32_t width, uint32_t height)
 		{
-			return std::floor(std::log2(glm::min(width, height))) + 1;
+			return (uint32_t)std::floor(std::log2(glm::min(width, height))) + 1;
 		}
 
 		inline uint32_t GetImageMemorySize(ImageFormat format, uint32_t width, uint32_t height)
